split segment into multiple modules
